# perf: cache `RepositoryPool.package()` in `Provider`