--- a/src/poetry/repositories/legacy_repository.py
+++ b/src/poetry/repositories/legacy_repository.py
@@ -46,9 +46,7 @@ class LegacyRepository(HTTPRepository):
             pool_size=pool_size,
         )
 
-    def package(
-        self, name: str, version: Version, extras: list[str] | None = None
-    ) -> Package:
+    def package(self, name: str, version: Version) -> Package:
         """
         Retrieve the release information.
 
@@ -65,7 +63,7 @@ class LegacyRepository(HTTPRepository):
 
             return self._packages[index]
         except ValueError:
-            package = super().package(name, version, extras)
+            package = super().package(name, version)
             package._source_type = "legacy"
             package._source_url = self._url
             package._source_reference = self.name