# Pull Request Check List

<!-- This is just a reminder about the most common mistakes. Please make sure that you tick all *appropriate* boxes.  But please read our [contribution guide](https://python-poetry.org/docs/contributing/) at least once, it will save you unnecessary review cycles! -->

- [ ] Added **tests** for changed code.
- [ ] Updated **documentation** for changed code.

This should be safe because we do not change the `RepositoryPool` while resolving and the returned package is used read-only.

Testing with the pyproject.toml from https://github.com/python-poetry/poetry/issues/9956#issuecomment-2582451109:

poetry version|lock (--regenerate)|lock (--no-update)
---|---|---
main (with python-poetry/poetry-core#828)|73 s|11.5 s
PR|52 s|5.7 s

## Summary by Sourcery

Cache the `RepositoryPool.package()` method in the `Provider` class to improve performance during dependency resolution.

Enhancements:
- Cache package lookups to improve performance.

Tests:
- Updated tests for the changes.